Rename to ExternalState uniformly
