+++ docs/site/src/content/docs/plugins/multiplatform/partial-match-fingerprint.mdx
@@ -65,7 +65,7 @@ plugin.get(assetId: "test")
 
 ## Partial Match Registry
 
-The `@player-ui/partial-match-registry` package provides a `Registry` that functions as a fuzzy-match dictionary. It's used by this plugin and others like the [External Action Plugin](/plugins/multiplatform/external-action/).
+The `@player-ui/partial-match-registry` package provides a `Registry` that functions as a fuzzy-match dictionary. It's used by this plugin and others like the [External State Plugin](/plugins/multiplatform/external-state/).
 
 The `Registry` maps `match`es (typically objects) to `value`s and retrieves the best value for a given query.
