+++ ios/demo/BUILD.bazel
@@ -17,8 +17,8 @@ swift_library(
         "//plugins/common-types/ios:PlayerUICommonTypesPlugin",
         "//plugins/common-expressions/ios:PlayerUICommonExpressionsPlugin",
         "//plugins/expression/ios:PlayerUIExpressionPlugin",
-        "//plugins/external-action/ios:PlayerUIExternalActionPlugin",
-        "//plugins/external-action/swiftui:PlayerUIExternalActionViewModifierPlugin",
+        "//plugins/external-state/ios:PlayerUIExternalStatePlugin",
+        "//plugins/external-state/swiftui:PlayerUIExternalStateViewModifierPlugin",
         "//plugins/metrics/swiftui:PlayerUIMetricsPlugin",
         "//plugins/console-logger/ios:PlayerUIPrintLoggerPlugin",
         "//plugins/pending-transaction/swiftui:PlayerUISwiftUIPendingTransactionPlugin",
