--- plugins/external-action/core/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@player-ui/external-action-plugin",
-  "version": "0.0.0-PLACEHOLDER",
-  "main": "src/index.ts",
-  "dependencies": {
-    "@player-ui/partial-match-registry": "workspace:*"
-  },
-  "peerDependencies": {
-    "@player-ui/player": "workspace:^"
-  }
-}
