--- plugins/external-action/core/src/symbols.ts
@@ -1,4 +1,0 @@
-// We prefix with "@player-ui" to avoid conflicts with symbols from other packages
-export const ExternalActionPluginSymbol: symbol = Symbol.for(
-  "@player-ui/ExternalActionPlugin",
-);
