--- plugins/external-action/ios/BUILD.bazel
@@ -1,6 +1,0 @@
-load("//tools/ios:util.bzl", "ios_plugin")
-
-ios_plugin(
-    name = "ExternalActionPlugin",
-    resources = ["//plugins/external-action/core:core_native_bundle"]
-)
