--- plugins/external-action/jvm/README.md
@@ -1,5 +1,0 @@
-# External Action Plugin
-
-## Gradle Dependency
-
-`implementation("com.intuit.playerui.plugins:external-action:$VERSION")`
