--- plugins/external-action/swiftui/BUILD.bazel
@@ -1,13 +1,0 @@
-load("//tools/ios:util.bzl", "swiftui_plugin")
-
-swiftui_plugin(
-    name = "ExternalActionViewModifierPlugin",
-    resources = [],
-    deps = [
-        "//ios/swiftui:PlayerUISwiftUI",
-        "//plugins/external-action/ios:PlayerUIExternalActionPlugin"
-    ],
-    test_deps = [
-        "//plugins/reference-assets/swiftui:PlayerUIReferenceAssets",
-    ]
-)
