+++ plugins/external-state/core/BUILD
@@ -1,0 +1,21 @@
+load("@rules_player//javascript:defs.bzl", "js_pipeline")
+load("@npm//:defs.bzl", "npm_link_all_packages")
+load("//tools:defs.bzl", "NATIVE_BUILD_DEPS", "tsup_config", "vitest_config")
+
+npm_link_all_packages(name = "node_modules")
+
+tsup_config(name = "tsup_config")
+
+vitest_config(name = "vitest_config")
+
+js_pipeline(
+    package_name = "@player-ui/external-state-plugin",
+    build_deps = NATIVE_BUILD_DEPS,
+    native_bundle = "ExternalStatePlugin",
+    peer_deps = [
+        ":node_modules/@player-ui/player",
+    ],
+    deps = [
+        ":node_modules/@player-ui/partial-match-registry",
+    ],
+)
