+++ plugins/external-state/core/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@player-ui/external-state-plugin",
+  "version": "0.0.0-PLACEHOLDER",
+  "main": "src/index.ts",
+  "dependencies": {
+    "@player-ui/partial-match-registry": "workspace:*"
+  },
+  "peerDependencies": {
+    "@player-ui/player": "workspace:^"
+  }
+}
