+++ plugins/external-state/core/src/symbols.ts
@@ -1,0 +1,4 @@
+// We prefix with "@player-ui" to avoid conflicts with symbols from other packages
+export const ExternalStatePluginSymbol: symbol = Symbol.for(
+  "@player-ui/ExternalStatePlugin",
+);
