+++ plugins/external-state/jvm/README.md
@@ -1,0 +1,5 @@
+# External State Plugin
+
+## Gradle Dependency
+
+`implementation("com.intuit.playerui.plugins:external-state:$VERSION")`
