+++ plugins/external-state/mocks/BUILD
