+++ plugins/external-state/swiftui/BUILD.bazel
@@ -1,0 +1,13 @@
+load("//tools/ios:util.bzl", "swiftui_plugin")
+
+swiftui_plugin(
+    name = "ExternalStateViewModifierPlugin",
+    resources = [],
+    deps = [
+        "//ios/swiftui:PlayerUISwiftUI",
+        "//plugins/external-state/ios:PlayerUIExternalStatePlugin"
+    ],
+    test_deps = [
+        "//plugins/reference-assets/swiftui:PlayerUIReferenceAssets",
+    ]
+)
