--- plugins/reference-assets/swiftui/UITests/ExternalActionUITests.swift
@@ -1,15 +1,0 @@
-import Foundation
-import XCTest
-
-class ExternalActionUITests: BaseTestCase {
-    override func navigateToAssetCollection() {
-        app.otherElements.buttons["Plugins + Managed Player"].firstMatch.tap()
-    }
-
-    func testExternalAction() {
-        openFlow("external-action external-action")
-        let alert = app.alerts["FlowCompleted"]
-        waitFor(alert)
-        XCTAssertTrue(alert.staticTexts.element(boundBy: 1).label.contains("FWD"))
-    }
-}
